move auto restart test in engine_integ_test.go
